+++ packages/transport-webbluetooth/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dg-kit/transport-webbluetooth
 
+## 1.2.1
+
+### Patch Changes
+
+- @dg-kit/core@1.2.1
+- @dg-kit/protocol@1.2.1
+
 ## 1.2.0
 
 ### Patch Changes